requirements.txt
snscrape
requests
beautifulsoup4
python-dotenv
certifi
